refactor(userlist): extract SwipeToDelete reusable component
Replace duplicate SwipeableListItem and SwipeableUserListItem private
composables with a shared SwipeToDelete component mirroring SwipeToAddBox.
